requirements.txt
apache-airflow
apache-airflow-providers-apache-spark
pyspark==3.5.4
boto3
apache-airflow-providers-common-compat
s3fs